renamed custom-components folder to custom-styles
