+++ src/components/custom-styles/ColoredText.ts
