+++ src/components/custom-styles/StyledSlider.tsx
